Refactor TOSA Legalizations to allow a legalization pass, to selectively apply the DECL_CONVERT_OP(*) allowing to choose whether to legalize the OP to TOSA on OP by OP case.
