+++ tensorflow/compiler/mlir/tosa/transforms/legalize_tfl.h
@@ -1,0 +1,143 @@
+/* Copyright 2020 The TensorFlow Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+
+// Legalize TensorFlow Lite to TOSA
+
+#include <climits>
+#include <cstddef>
+#include <cstdint>
+#include <fstream>
+#include <iterator>
+#include <numeric>
+#include <unordered_set>
+
+#include "mlir/Dialect/Quant/QuantOps.h"  // from @llvm-project
+#include "mlir/Dialect/Quant/QuantTypes.h"  // from @llvm-project
+#include "mlir/Dialect/Tosa/IR/TosaOps.h"  // from @llvm-project
+#include "mlir/IR/BuiltinAttributes.h"  // from @llvm-project
+#include "mlir/IR/BuiltinTypes.h"  // from @llvm-project
+#include "mlir/IR/PatternMatch.h"  // from @llvm-project
+#include "mlir/Support/LLVM.h"  // from @llvm-project
+#include "mlir/Transforms/DialectConversion.h"  // from @llvm-project
+#include "tensorflow/compiler/mlir/lite/ir/tfl_ops.h"
+#include "tensorflow/compiler/mlir/tosa/transforms/legalize_common.h"
+#include "tensorflow/compiler/mlir/tosa/transforms/legalize_utils.h"
+#include "tensorflow/compiler/mlir/tosa/transforms/passes.h"
+
+#define PASS_NAME "tosa-legalize-tfl"
+#define DEBUG_TYPE PASS_NAME
+#define HARDSWISH_EXPLICIT_RESCALING false
+
+namespace mlir {
+namespace tosa {
+namespace {
+
+#include "tensorflow/compiler/mlir/tosa/transforms/tfl_legalize_patterns.inc"
+
+#define DECL_CONVERT_OP(tfl_op)                                              \
+  struct ConvertTFL##tfl_op##Op : public RewritePattern {                    \
+    explicit ConvertTFL##tfl_op##Op(MLIRContext* context)                    \
+        : RewritePattern(TFL::tfl_op##Op::getOperationName(), 1, context) {} \
+    LogicalResult matchAndRewrite(Operation* op,                             \
+                                  PatternRewriter& rewriter) const override; \
+  }
+DECL_CONVERT_OP(Relu);
+DECL_CONVERT_OP(Relu6);
+DECL_CONVERT_OP(Equal);
+DECL_CONVERT_OP(NotEqual);
+DECL_CONVERT_OP(Greater);
+DECL_CONVERT_OP(GreaterEqual);
+DECL_CONVERT_OP(Add);
+DECL_CONVERT_OP(Sub);
+DECL_CONVERT_OP(Mul);
+DECL_CONVERT_OP(Square);
+DECL_CONVERT_OP(SquaredDifference);
+DECL_CONVERT_OP(Round);
+DECL_CONVERT_OP(Div);
+DECL_CONVERT_OP(Maximum);
+DECL_CONVERT_OP(Minimum);
+DECL_CONVERT_OP(FloorMod);
+DECL_CONVERT_OP(FloorDiv);
+DECL_CONVERT_OP(AddN);
+DECL_CONVERT_OP(AveragePool2D);
+DECL_CONVERT_OP(MaxPool2D);
+DECL_CONVERT_OP(Concatenation);
+DECL_CONVERT_OP(Reshape);
+DECL_CONVERT_OP(Rank);
+DECL_CONVERT_OP(Shape);
+DECL_CONVERT_OP(ExpandDims);
+DECL_CONVERT_OP(Squeeze);
+DECL_CONVERT_OP(Fill);
+DECL_CONVERT_OP(Elu);
+DECL_CONVERT_OP(Softmax);
+DECL_CONVERT_OP(LogSoftmax);
+DECL_CONVERT_OP(Sqrt);
+DECL_CONVERT_OP(L2Normalization);
+DECL_CONVERT_OP(ReduceAny);
+DECL_CONVERT_OP(ReduceMax);
+DECL_CONVERT_OP(ReduceMin);
+DECL_CONVERT_OP(Mean);
+DECL_CONVERT_OP(ReduceProd);
+DECL_CONVERT_OP(Sum);
+DECL_CONVERT_OP(Conv2D);
+DECL_CONVERT_OP(TransposeConv);
+DECL_CONVERT_OP(DepthwiseConv2D);
+DECL_CONVERT_OP(FullyConnected);
+DECL_CONVERT_OP(BatchMatMul);
+DECL_CONVERT_OP(Split);
+DECL_CONVERT_OP(SplitV);
+DECL_CONVERT_OP(Pack);
+DECL_CONVERT_OP(Unpack);
+DECL_CONVERT_OP(Transpose);
+DECL_CONVERT_OP(Tile);
+DECL_CONVERT_OP(Slice);
+DECL_CONVERT_OP(StridedSlice);
+DECL_CONVERT_OP(HardSwish);
+DECL_CONVERT_OP(ZerosLike);
+DECL_CONVERT_OP(Less);
+DECL_CONVERT_OP(LessEqual);
+DECL_CONVERT_OP(Pad);
+DECL_CONVERT_OP(PadV2);
+DECL_CONVERT_OP(ResizeBilinear);
+DECL_CONVERT_OP(ResizeNearestNeighbor);
+DECL_CONVERT_OP(Select);
+DECL_CONVERT_OP(SelectV2);
+DECL_CONVERT_OP(SpaceToBatchNd);
+DECL_CONVERT_OP(BatchToSpaceNd);
+DECL_CONVERT_OP(SpaceToDepth);
+DECL_CONVERT_OP(DepthToSpace);
+DECL_CONVERT_OP(Logistic);
+DECL_CONVERT_OP(Tanh);
+DECL_CONVERT_OP(PRelu);
+DECL_CONVERT_OP(LeakyRelu);
+DECL_CONVERT_OP(Neg);
+DECL_CONVERT_OP(Yield);
+DECL_CONVERT_OP(Custom);
+DECL_CONVERT_OP(ReverseV2);
+DECL_CONVERT_OP(Quantize);
+DECL_CONVERT_OP(Dequantize);
+DECL_CONVERT_OP(Const);
+DECL_CONVERT_OP(QConst);
+DECL_CONVERT_OP(Gather);
+DECL_CONVERT_OP(GatherNd);
+DECL_CONVERT_OP(SparseToDense);
+DECL_CONVERT_OP(OneHot);
+DECL_CONVERT_OP(ArgMax);
+DECL_CONVERT_OP(FakeQuant);
+#undef DECL_CONVERT_OP
+
+}  // namespaces
+}  // namespace tosa
+}  // namespace mlir
